review comments

+++ include/battle_dynamax.h
@@ -14,7 +14,7 @@ bool32 IsMoveBlockedByMaxGuard(u32 move);
 bool32 IsMoveBlockedByDynamax(u32 move);
 
 u16 GetMaxMove(u32 battler, u32 baseMove);
-u8 GetMaxMovePower(u32 move);
+u32 GetMaxMovePower(u32 move);
 bool32 IsMaxMove(u32 move);
 void ChooseDamageNonTypesString(u8 type);
